renamed backend to compiler_interface
